first cut at metrics_cache table which replaces metrics_index table
The number of reads performed during aggregation grow linearly in proportion to
the number of schedules with data to be aggregated. In the worst case scenario
when 1 hr and 6 hr data are aggregated along with raw data, the number of reads
is 3 * N where N is the number of schedules. If there are a million schedules
with data, then a million reads have to be performed just to aggregate the raw
data.

The new metrics_cache table allows us to dramatically reduce the number of
reads. When raw data for example is stored it is now written to both the
raw_data table as well as the metrics_cache table. A single partition, i.e.,
row, in metrics_cache contains data for some configurable number of schedules.
Right now it defaults to 100. Each partition in metrics_cache contains data
only for a given time slice, e.g., 18:00 to 19:00. Once aggregation has run for
that time slice, the partition is deleted.

If we have a million schedules with raw data and 100 schedules are being stored
per partition in metrics_cache, then the number of reads drops from 1 million
to 10,000. That is a 99% reduction in the number of reads. And this reduction
will not come with a sharp increase in memory utilization.

+++ modules/helpers/metrics-simulator/src/main/java/org/rhq/metrics/simulator/Simulator.java
@@ -108,7 +108,7 @@ public class Simulator implements ShutdownManager {
         Random random = new Random();
         long timestamp = plan.getDateTimeService().nowInMillis();
         long endOfSimulation = timestamp + 24L * 60 * 60 * 1000 * plan.getSimulationTime();
-        long numberOfMetrics = plan.getBatchSize() * plan.getNumMeasurementCollectors();
+        int numberOfMetrics = plan.getBatchSize() * plan.getNumMeasurementCollectors();
 
         Set<MeasurementDataNumeric> data = new HashSet<MeasurementDataNumeric>(plan.getBatchSize());
